+++ mobile/android/android-components/docs/changelog.md
@@ -33,6 +33,9 @@ permalink: /changelog/
 * **browser-icons**
   * Fixed possible `NullPointerException` when disk cache is written to concurrently.
 
+* **lib-crash**
+  * Crash reports sent to Sentry now contain optional environment information, if a parameter is passed.
+  
 # 0.55.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.54.0...v0.55.0)